kubernetes-config/helm-charts/app/src/requirements.txt
flask
psycopg2-binary
azure-identity
